+++ packages/react/src/components/Collapsible/index.tsx
@@ -1,0 +1,51 @@
+import React, { ComponentProps } from 'react'
+import * as CollapsiblePrimitive from '@radix-ui/react-collapsible'
+import { CaretDown } from 'phosphor-react'
+
+import {
+  CollapsibleRoot,
+  StyledTrigger,
+  Flex as FlexContainer,
+  CollapsibleText as Text,
+  Repository as ContentRepository,
+  StyledContent,
+} from './styles'
+import { CSS } from '../../styles'
+
+type CollapsiblePrimitiveProps = ComponentProps<
+  typeof CollapsiblePrimitive.Root
+>
+type CollapsibleProps = CollapsiblePrimitiveProps & { css?: CSS }
+
+// eslint-disable-next-line react/display-name
+export const Collapsible = React.forwardRef<
+  React.ElementRef<typeof CollapsibleRoot>,
+  CollapsibleProps
+>(({ children, ...props }, forwardedRef) => (
+  <CollapsibleRoot ref={forwardedRef} {...props}>
+    {children}
+  </CollapsibleRoot>
+))
+
+type CollapsibleTriggerPrimitiveProps = ComponentProps<
+  typeof CollapsiblePrimitive.Trigger
+>
+type CollapsibleTriggerProps = CollapsibleTriggerPrimitiveProps & { css?: CSS }
+
+// eslint-disable-next-line react/display-name
+export const CollapsibleTrigger = React.forwardRef<
+  React.ElementRef<typeof StyledTrigger>,
+  CollapsibleTriggerProps
+>(({ children, ...props }, forwardedRef) => (
+  <StyledTrigger {...props} ref={forwardedRef}>
+    {children}
+    <CaretDown />
+  </StyledTrigger>
+))
+
+export const Flex = FlexContainer
+export const CollapsibleText = Text
+export const Repository = ContentRepository
+export const CollapsibleContent = StyledContent
+
+Collapsible.displayName = 'Collapsible'